+++ src/elements/index.js
@@ -1,6 +1,9 @@
 const Action = require('./action');
+const Annotation = require('./annotation');
 const Api = require('./api');
+const BaseElement = require('./base');
 const Example = require('./example');
+const Header = require('./header');
 const Request = require('./request');
 const Resource = require('./resource');
 const Response = require('./response');
@@ -10,8 +13,11 @@ const UriParameter = require('./uri-parameter');
 
 module.exports = {
   Action,
+  Annotation,
   Api,
+  BaseElement,
   Example,
+  Header,
   Request,
   Resource,
   Response,